Add Test Code & Refactored Code
1. 시나리오 대로 Test Code 작성.
   - Team과 Account에 OneToOne으로 추가될 연관관계 때문에 Team 관련
테스트 코드는 추후에 작성 예정.

2. 기존에 Account의 Name으로 만들었던 JWT를 Email로 만들도록 변경.
   - Name은 중복이 될 수 있으므로 Email로 토큰을 만들되, Email 중복
방지가 요구 됨.

3. 기존에 UserDetails를 상속 받던 Account 변경.
   - Entity클래스인 Account에 DB와는 불필요한 오버라이드 메서드들이
생김.
   - UserDetails를 상속받는 AccountAdapter.class 생성.
     - 결국 우리는 loadUserByUsername에서 Account객체의 값을 가지는
UserDetails를 반환해야함.
     - 지금까지는 UserDetails를 상속받는 Account를 반환.
     - 수정된 코드는 UserDetails를 상속받고, Account의 값을 가지는
AccountAdapter를 반환.
   - 그래서 AccountDetailsService.loadUserByUsername에서 new
AccountAdapter(account) 를 반환 함.

4. AccountDetailsService 위치를 account 패키지 안으로 변경.
   - Account와 관련된 로직이므로 위치 이동.
   - 기존에는 DB에서 찾은 account를 반환했으나, 현재는 new
AccountAdapter(account)를 반환 함.

5. CurrentAccount 커스텀 애노테이션 생성
   - 현재 로그인한 유저를 불러오기 위한 애노테이션 추가

6. 불필요한 공백 및 import문 삭제

+++ src/main/java/com/ksu/soccerserver/account/AccountController.java
@@ -1,30 +1,24 @@
 package com.ksu.soccerserver.account;
 
-
 import com.ksu.soccerserver.config.JwtTokenProvider;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-
 import java.util.Collections;
-import java.util.Map;
-
 import com.ksu.soccerserver.team.Team;
 import com.ksu.soccerserver.team.TeamRepository;
-import lombok.RequiredArgsConstructor;
-import org.apache.catalina.webresources.JarResourceRoot;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
 
-
 @RequestMapping("/api/accounts")
 @RequiredArgsConstructor
 @RestController
 public class AccountController {
+
     private final PasswordEncoder passwordEncoder;
     private final JwtTokenProvider jwtTokenProvider;
     private final AccountRepository accountRepository;
@@ -53,23 +47,31 @@ public class AccountController {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "잘못된 비밀번호입니다.");
         }
 
-        return new ResponseEntity<> (jwtTokenProvider.createToken(member.getName(), member.getRoles()), HttpStatus.OK);
+        return new ResponseEntity<> (jwtTokenProvider.createToken(member.getEmail(), member.getRoles()), HttpStatus.OK);
     }
 
 
 
     // 회원정보 출력
     @GetMapping("/{accountId}")
-    public ResponseEntity<?> getAccount(@PathVariable Long accountId){
+    public ResponseEntity<?> getAccount(@PathVariable Long accountId, @CurrentAccount Account currentAccount){
         Account findAccount = accountRepository.findById(accountId).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "존재하지 않는 유저입니다."));
+
+        if(!currentAccount.getId().equals(findAccount.getId())) {
+            return new ResponseEntity<>("권한이 없습니다.", HttpStatus.BAD_REQUEST);
+        }
 
         return new ResponseEntity<>(findAccount, HttpStatus.OK);
     }
 
     // 회원정보 수정
     @PutMapping("/{accountId}")
-    public ResponseEntity<?> postAccount(@PathVariable Long accountId, @RequestBody Account account) {
+    public ResponseEntity<?> postAccount(@PathVariable Long accountId, @RequestBody Account account, @CurrentAccount Account currentAccount) {
         Account findAccount = accountRepository.findById(accountId).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "존재하지 않는 유저입니다."));
+
+        if(!currentAccount.getId().equals(findAccount.getId())) {
+            return new ResponseEntity<>("권한이 없습니다.", HttpStatus.BAD_REQUEST);
+        }
 
         findAccount.updateMyInfo(account.getName());
         accountRepository.save(findAccount);
@@ -94,8 +96,12 @@ public class AccountController {
 
     // 회원 탈퇴
     @DeleteMapping("/{accountId}")
-    public ResponseEntity<?> deleteAccount(@PathVariable Long accountId){
+    public ResponseEntity<?> deleteAccount(@PathVariable Long accountId, @CurrentAccount Account currentAccount){
         Account findAccount = accountRepository.findById(accountId).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "존재하지 않는 유저입니다."));
+
+        if(!currentAccount.getId().equals(findAccount.getId())) {
+            return new ResponseEntity<>("권한이 없습니다.", HttpStatus.BAD_REQUEST);
+        }
 
         accountRepository.delete(findAccount);
